Updating Signup and Validation Scripts
New JS validation for signup that wasn't included in my last commit.
Functions are modular so they can be very quickly and easily adapted to
work on the sign-in page as well.

Brought the Signup.html file up to date with the appropriate divs for
form validation error messages, css for those divs, and appropriate JS
references. I also reformatted the code so all <tags> had closing
</tags> and indentation was proper. The <nav> portion may be redundant,
but I will be asking Nick about that.